requirements.txt
yfinance
pandas
sqlalchemy
psycopg2-binary
python-dotenv
streamlit
plotly
numpy==1.23.5